Fix code reviewer suggestion
